Rework Collection
Extending an array in Typescript doesnt work as you might expect.
Since Javascript is prototypal the object returned from mutators
of an extended class are of type Array rather than the extended
type. To ensure predictability we dont try and work around this.
All collection methods always return type Array. The caller will
have to cast the result to the desired type.